feat: overhaul cli command structure (#2)
This commit introduces a command structure which will support the addition
of other database types (by adding `Introspectors`) and other diagram
markup languages (by adding `Renderers`). In the new structure, the
subcommand choses the which database type will be used (currently
`sqlite` or `libSQL`) and the `--format` flag will determine the output
format (currently only `d2` which is also the default). A complete list
of changes:

- Add `libsql`/`sqlite` subcommands
- `—format` flag for diagram format (currently only `d2` suppported)
- `—exclude-table` flag to exlude tables from diagrams
- `—generated-tables` flag to include generated tables
- Improve parsing/validation of libsql/sqlite url argument
- Improve flag summaries
- Improve flag placeholders

---------

Signed-off-by: Luca Schultz <github@lucaschultz.com>